test(core): reorganize test file structure
