+++ packages/core/__tests__/playwright/stories/exam/DNDAnswerContainer.story.tsx
@@ -1,11 +1,11 @@
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
 import React from 'react'
 import { MasteringResult } from '@digabi/exam-engine-mastering'
-import { renderChildNodes } from '../../../src/components/exam/Exam'
-import { DNDAnswerContainer } from '../../../src/components/exam/DNDAnswerContainer'
-import { ComponentWrapper } from '../utils/ComponentWrapper'
-import parseExam from '../../../src/parser/parseExam'
-import { queryAll } from '../../../src/dom-utils'
+import { renderChildNodes } from '../../../../src/components/exam/Exam'
+import { DNDAnswerContainer } from '../../../../src/components/exam/DNDAnswerContainer'
+import { ExamComponentWrapper } from '../../utils/ExamComponentWrapper'
+import parseExam from '../../../../src/parser/parseExam'
+import { queryAll } from '../../../../src/dom-utils'
 
 export type answerMediaType = 'text' | 'image'
 
@@ -18,7 +18,7 @@ export const DNDAnswerContainerStory: React.FC<DNDAnswerContainerStoryProps> = (
   const doc = parseExam(masteredExam.xml, true)
   const [textAnswers, imageAnswers] = queryAll(doc.documentElement, 'dnd-answer-container')
   return (
-    <ComponentWrapper
+    <ExamComponentWrapper
       doc={doc}
       resolveAttachment={(filename: string) => `/${masteredExam.examCode}/attachments/${filename}`}
     >
@@ -26,6 +26,6 @@ export const DNDAnswerContainerStory: React.FC<DNDAnswerContainerStoryProps> = (
         element={answerMediaType === 'text' ? textAnswers : imageAnswers}
         renderChildNodes={renderChildNodes}
       ></DNDAnswerContainer>
-    </ComponentWrapper>
+    </ExamComponentWrapper>
   )
 }
